+++ tests/testthat/helper.R
@@ -38,8 +38,11 @@ scoped_temporary_thing <- function(dir = file_temp(pattern = pattern),
       )
     }
   } else {
-    withr::defer(proj_set(old_project, force = TRUE), envir = env)
-    withr::defer(fs::dir_delete(dir), envir = env)
+    withr::defer({
+      proj_set(old_project, force = TRUE)
+      setwd(old_project)
+      fs::dir_delete(dir)
+    }, envir = env)
   }
 
   withr::local_options(list(usethis.quiet = TRUE))
@@ -48,6 +51,8 @@ scoped_temporary_thing <- function(dir = file_temp(pattern = pattern),
     package = create_package(dir, rstudio = rstudio, open = FALSE),
     project = create_project(dir, rstudio = rstudio, open = FALSE)
   )
+  proj_set(dir)
+  setwd(dir)
   invisible(dir)
 }